requirements.txt
discord.py
python-dotenv
aiohttp
async-lru
